+++ src/Instance/ItemsList/ItemsListContainer.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { act } from 'react-dom/test-utils';
+import { act } from '@folio/jest-config-stripes/testing-library/react';
 
 import '../../../test/jest/__mock__';