+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add permission for "Share local instance" option on Member tenants. Refs UIIN-3140.
 * React v19: refactor away from default props for functional components. Refs UIIN-2890.
 * User can edit Source consortium "Holdings sources" in member tenant but not in Consortia manager. Refs UIIN-3147.
+* React 19: refactor away from react-dom/test-utils. Refs UIIN-2888.
 
 ## [12.0.5] (IN PROGRESS)